Theme extensions

Some functions in ggh4x are using extensions to the theme system. These extended theme argument are listed below, along with what elements they are expected to be, and in what function(s) they are used.

Arguments

ggh4x.facet.nestline

An element_line() used as the parent for the nest_line argument in facet_nested() and facet_nested_wrap(). Inherits directly from the 'line' theme element.

ggh4x.axis.nestline, ggh4x.axis.nestline.x, ggh4x.axis.nestline.y

An element_line() used as the line to separate different layers of labels in guide_axis_nested(). Inherits from the 'axis.ticks' theme element.

ggh4x.axis.nesttext.x, ggh4x.axis.nesttext.y

An element_text() used to differentiate text higher in the hierarchy from the axis labels directly next to the axis line in guide_axis_nested(). Inherits from the 'axis.text.x' and 'axis.text.y' theme elements respectively.

ggh4x.axis.ticks.length.minor

A rel() object used to set the size of minor tick marks relative to the regular tick marks. This is used in the guide_axis_minor() and guide_axis_logticks() functions. Defaults to rel(2/3).

ggh4x.axis.ticks.length.mini

A rel() object used to set the size of the smallest tick marks relative to regular tick marks. This is only used in the guide_axis_logticks() function. Defaults to rel(1/3).